Updates the packrat library: qdapTools
The packrat library was updated. Several packages were affected but the
most notable change made was in 'qdapTools'. This was done to address the
bug noticed in R 3.4 series that results in a namespace clash between
'qdap' and 'openssl'. Fix was introduced by the package author in the dev
version of the package obtained from GitHub (v. 1.3.4).