Initial commit on new files
The user UI and server code have been separated to simplify code maintenance.
Also, new condition handling was introduced, and the app no longer stops executing when any of the components are not functioning properly. For instance, a recent problem with qdap::polarity (Namespace clash with openssl) has resurfaced, possibly following an update, and will be addressed at a later time.